+++ App.js
@@ -1,9 +1,59 @@
-import React from 'react';
-import { Navigation } from './Navigation';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from "@react-navigation/native-stack"
+import { Login, Signup, Welcome, Welcome2, SignupEmpresa, Form } from "./screens";
+
+const Stack = createNativeStackNavigator();
 
 export default function App() {
-return(
-  <Navigation />
-)
-
+  
+  return (
+    <NavigationContainer>
+      <Stack.Navigator
+        initialRouteName='Welcome'
+      >
+        <Stack.Screen
+          name="Welcome"
+          component={Welcome}
+          options={{
+            headerShown: false
+          }}
+        />
+        <Stack.Screen
+          name="Welcome2"
+          component={Welcome2}
+          options={{
+            headerShown: false
+          }}
+        />
+        <Stack.Screen
+          name="Login"
+          component={Login}
+          options={{
+            headerShown: false
+          }}
+        />
+        <Stack.Screen
+          name="Signup"
+          component={Signup}
+          options={{
+            headerShown: false
+          }}
+        />
+        <Stack.Screen
+          name="Form"
+          component={Form}
+          options={{
+            headerShown: false
+          }}
+        />
+        <Stack.Screen
+        name="SignupEmpresa"
+        component={SignupEmpresa}
+        options={{
+          headerShown: false
+        }}
+        />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
 }
