+++ package.json
@@ -12,7 +12,6 @@
     "@expo/webpack-config": "^18.0.1",
     "@react-native-google-signin/google-signin": "^7.2.2",
     "@react-native-picker/picker": "^2.4.10",
-    "@react-navigation/bottom-tabs": "^6.5.7",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
     "expo": "~48.0.15",
