+++ screens/Form.js
@@ -649,7 +649,6 @@ const Signup = ({ navigation }) => {
 
                 <Button
                     title="Agregar información al perfil"
-                    onPress={() => navigation.navigate("Home")}
                     filled
                     style={{
                         marginTop: 18,
