+++ screens/index.js
@@ -4,10 +4,6 @@ import Signup from "./Signup";
 import Welcome2 from "./Welcome2";
 import SignupEmpresa from "./SignupEmpresa";
 import Form from "./Form"
-import Notificaciones from "./Notificaciones"
-import Home from "./HomeScreen";
-import Perfil from "./Perfil";
-
 
 export {
     Login,
@@ -15,8 +11,5 @@ export {
     Welcome2,
     Signup,
     SignupEmpresa,
-    Form,
-    Notificaciones,
-    Home,
-    Perfil,
+    Form
 }
